// src/mocks/types.ts
export type ProfileAvatarGroupType = {
    name: string
    avatar: string
}
// ...
export type ProjectsTabType = {
    hours: string
    tasks: string
    title: string
    budget: string
    client: string
    avatar: string
    members: string
    daysLeft: number
    comments: number
    deadline: string
    startDate: string
    totalTask: number
    budgetSpent: string
    description: string
    chipColor: ThemeColor
    completedTask: number
    avatarColor?: ThemeColor
    avatarGroup: ProfileAvatarGroupType[]
}